apps/ingestion/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
httpx==0.27.0
elasticsearch[async]==8.12.0
python-dotenv==1.0.1
celery>=5.3.6
redis>=5.0.1